pyproject.toml
[tool.poetry]
name = "imageupscaler"
version = "0.1.0"
description = "TensorFlow-based image super-resolution service with FastAPI and modern frontend"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"  # TensorFlow 2.16 & Pillow wheels unavailable for 3.13 yet
fastapi = "^0.111.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
pillow = "^10.3.0"
numpy = "^1.26.0"
protobuf = "^4.25.3"
python-multipart = "^0.0.9"
loguru = "^0.7.2"
pydantic = "^1.10.15"  # pin v1 for BaseSettings import

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
black = "^24.4.0"
pytest = "^8.2.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"

[tool.poetry.group.ml.dependencies]
tensorflow = "^2.16.1"  # install with: poetry install --with ml
tensorflow-metal = "^1.1.0"  # M1 Mac GPU acceleration
tensorflow-hub = "^0.16.0"  # Pre-trained model hub

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
